Move MainActorDebouncer into it’s own file
