+++ firefox-ios/firefox-ios-tests/Tests/ClientTests/Helpers/AccountSyncHandlerTests.swift
@@ -100,80 +100,6 @@ class AccountSyncHandlerTests: XCTestCase {
         wait(for: [expectation], timeout: 1.0)
         XCTAssertEqual(profile.storeAndSyncTabsCalled, 2)
     }
-
-    // MARK: - MainActorDebouncer
-
-    func testDebouncer_subSecondDelay_doesNotExecuteBeforeDelayElapses() {
-        let expectation = XCTestExpectation(description: "action does not execute before a sub-second delay elapses")
-        expectation.isInverted = true
-        let subject = MainActorDebouncer(delay: 0.5)
-
-        subject.call { expectation.fulfill() }
-
-        wait(for: [expectation], timeout: 0.2)
-        subject.cancel()
-    }
-
-    func testDebouncer_subSecondDelay_executesAfterDelayElapses() {
-        let expectation = XCTestExpectation(description: "action executes once a sub-second delay elapses")
-        let subject = MainActorDebouncer(delay: 0.1)
-
-        subject.call { expectation.fulfill() }
-
-        wait(for: [expectation], timeout: 1.0)
-    }
-
-    func testDebouncer_zeroDelay_executes() {
-        let expectation = XCTestExpectation(description: "action executes when the delay is zero")
-        let subject = MainActorDebouncer(delay: 0)
-
-        subject.call { expectation.fulfill() }
-
-        wait(for: [expectation], timeout: 1.0)
-    }
-
-    func testDebouncer_negativeDelay_executesWithoutTrapping() {
-        let expectation = XCTestExpectation(description: "action executes when the delay is negative")
-        let subject = MainActorDebouncer(delay: -1.0)
-
-        subject.call { expectation.fulfill() }
-
-        wait(for: [expectation], timeout: 1.0)
-    }
-
-    func testDebouncer_cancel_preventsPendingAction() {
-        let expectation = XCTestExpectation(description: "cancelled action does not execute")
-        expectation.isInverted = true
-        let subject = MainActorDebouncer(delay: 0.1)
-
-        subject.call { expectation.fulfill() }
-        subject.cancel()
-
-        wait(for: [expectation], timeout: 0.5)
-    }
-
-    func testDebouncer_callAfterCancel_executes() {
-        let expectation = XCTestExpectation(description: "action scheduled after a cancel executes")
-        let subject = MainActorDebouncer(delay: 0.1)
-
-        subject.call { XCTFail("the cancelled action should not execute") }
-        subject.cancel()
-        subject.call { expectation.fulfill() }
-
-        wait(for: [expectation], timeout: 1.0)
-    }
-
-    func testDebouncer_multipleCalls_executesOnlyTheMostRecentAction() {
-        let supersededAction = XCTestExpectation(description: "superseded action does not execute")
-        supersededAction.isInverted = true
-        let latestAction = XCTestExpectation(description: "most recent action executes")
-        let subject = MainActorDebouncer(delay: 0.1)
-
-        subject.call { supersededAction.fulfill() }
-        subject.call { latestAction.fulfill() }
-
-        wait(for: [supersededAction, latestAction], timeout: 0.5)
-    }
 }
 
 // MARK: - Helper methods
